update props for record table; remove recordgroup prop
